Added send_email_script javascript file that will send the email and changed staff request page js to include this change
Added an email script that needs a login and password for the bot email
account. Then it needs someone to enter an email that will recieve the
bot's email. Currently, this code does not have a way to get the
information from the request page .js, but I have set up comments that
tell us how to send the information over.

In request_page.js, I deleted the old email script and i made it so it
calls the main function from the send_script javascript file. The
pathing has not been tested yet